Obstacles spawn at spawn points
Spawner spawns obstacles where spawn points are placed in the scene.
Obstacles move from the spawn point.
Obstacles spawn continuously between intervals.
Obstacles are destroyed after delay.
