--- Assets/YOUR_STUFF_HERE/PlayerDog.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a529ab19268f9ec408874fabf14c5872
