--- Assets/YOUR_STUFF_HERE/SausageDogManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 44aff5c2a226fa541972352b9a0d7689
