+++ Assets/YOUR_STUFF_HERE/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 00fae16d394ab4544939a7f868c6fcb6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
