+++ Assets/YOUR_STUFF_HERE/Scripts/ObstacleSpawner.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 9878bc60ba0186f40ae5f3e5f5ea7586
