+++ Assets/YOUR_STUFF_HERE/Scripts/PlayerDog.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a529ab19268f9ec408874fabf14c5872
