+++ Assets/YOUR_STUFF_HERE/Scripts/SausageDogManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 44aff5c2a226fa541972352b9a0d7689
